Refactor: Restrictire views
