+++ frontend/src/router.js
@@ -1,14 +1,14 @@
 import { createRouter, createWebHistory } from "vue-router";
 const HomeView = () => import("./views/Home.vue");
-const AdminHomeView = () => import( "./views/AdminHome.vue");
-const MenuView = () => import( "./views/MenuRender.vue");
-const AdminSettingsView = () => import( "./views/AdminSettings.vue");
-const VendorSettings = () => import( "./views/VendorSettings.vue");
-const VendorMenuManager = () => import( "./views/VendorMenuManager.vue");
-const AdminVendorsView = () => import( "./views/AdminVendors.vue");
-const AdminOrdersView = () => import( "./views/AdminOrders.vue");
-const AdminUsersView = () => import( "./views/AdminUsers.vue");
-const VendorItemManager = () => import( "./views/VendorItemManager.vue");
+const AdminHomeView = () => import( "./views/admin/AdminHome.vue");
+const MenuView = () => import( "./views/menu/MenuRender.vue");
+const AdminSettingsView = () => import( "./views/admin/AdminSettings.vue");
+const VendorSettings = () => import( "./views/admin/vendor/VendorSettings.vue");
+const VendorMenuManager = () => import( "./views/admin/vendor/VendorMenuManager.vue");
+const AdminVendorsView = () => import( "./views/admin/AdminVendors.vue");
+const AdminOrdersView = () => import( "./views/admin/AdminOrders.vue");
+const AdminUsersView = () => import( "./views/admin/AdminUsers.vue");
+const VendorItemManager = () => import( "./views/admin/vendor/VendorItemManager.vue");
 const VendorAdd = () => import( "./components/VendorAdd.vue");
 const NotFound = () => import( "./components/NotFound.vue");
 import { useAuth } from "@/stores/auth.js";
