+++ frontend/src/views/admin/AdminVendors.vue
